// frontend/app/types/api.ts
export interface Deviation {
    ticketId: string | null;
    type: string;
    description: string;
    createdAt: Date;
}